ground-truths-list/requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro~=0.19.6
kedro-datasets>=3.0; python_version >= "3.9"
kedro-datasets>=1.0; python_version < "3.9"
kedro-telemetry>=0.3.1
kedro-viz>=6.7.0
notebook
pytest~=7.2
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
ruff~=0.1.8
setuptools; python_version >= "3.12"
pyspark
openpyxl
google.generativeai
vertexai
rdflib
matplotlib
asyncio
aiohttp
